+++ CHANGES.md
@@ -74,6 +74,7 @@ Changes in 12.5
 - Add `_gradeScaleInputHtml()` & `_search()` private methods to Widget_letter_grade object in Widget.php
 - Use Select2 input + AJAX search when scale has more than 1003 grades in Widget.php
 - Change User-Agent to "RosarioSIS/1.0; +https://www.rosariosis.org/" in curl.php
+- Cache config values, avoid calling Config() function X times in Currency.fnc.php
 
 Changes in 12.4.3
 -----------------